refactor(cli/claude_hook): convert to a directory; lift the test module to tests.rs
claude_hook.rs was 1103 lines; ~523 were the trailing #[cfg(test)] module. Lift it
into claude_hook/tests.rs, dropping production to 562 lines. Pure move.
